Add ROS service to take new Spot lease (#219)
To support shared autonomy with multiple clients (e.g. `bosdyn.client.RobotCommandClient`), we need a way to ask `SpotROS` to take a new lease.

The underlying SpotWrapper exposes a `getLease` interface, which either takes or acquires the lease, depending on the configuration for the driver. This PR uses the underlying interface of SpotWrapper directly to expose a ROS service that takes the lease.

For example, a teleop agent may have taken control of Spot via `lease_client.take()`, and a skill commanding spot through `SpotROS` will need to ask `SpotROS` to generate a new lease when teleop is done in order to be able to control Spot again (because`SpotROS` would be running with `use_take_lease=True` and `get_lease_on_action=False`).  

internal ticket SW-653 tracks the spot_wrapper changes to avoid private interface access in here.