[SW-632] Check config file path exists in spot_ros2 (#246)
## Change Overview
The goal of this PR is to check if the `config_file` launch argument for `spot_driver.launch.py` has been set, and if so, checks if it is a valid path. If this file is not valid, an exception is thrown. This happens before any nodes get launched so that there is no chance of partially launched "zombie nodes". 
## Testing Done
I tested by running the spot driver with a physical Spot, passing in a nonexistent config file path. In this case the driver exits before any nodes get launched. If the config file was a real file, or if no config file was passed in (for example, if the user chooses to set parameters through environment variables instead) the behavior is unchanged. 
Example of the failure state:
```
$ ros2 launch spot_driver spot_driver.launch.py config_file:=notarealfile.yaml
[INFO] [launch]: All log files can be found below /home/khughes/.ros/log/2024-01-31-10-17-41-390265-khughes-LT-900652
[INFO] [launch]: Default logging verbosity is set to INFO
[ERROR] [launch]: Caught exception in launch (see debug for traceback): Configuration file 'notarealfile.yaml' does not exist!
$ 
```