[SW-642] Make it obvious when spot is running out of battery (#252)
Created a new `spot_alerts` node to monitor stats and throw a pop-up alert warning the user. It currently only monitors the battery states and throws a warning up if the percentage is <= 10%.
Also added the `low_battery` parameter flag to prevent constant pop up spamming.

Tested on a spot with low battery.