[SW-794] hotfix for spot driver launchfile not launching (#282)
## Change Overview

Current `spot_driver.launch.py` launchfile on main does not work, fails with
`executable 'kinematic_node' not found on the libexec directory`
and, when this was fixed,
`executable 'spot_robot_state_publisher_node' not found on the libexec directory`

These errors were introduced with #209 which implemented the robot state publisher in C++ and also changed the name of some nodes. This just updates to the correct node names.

## Testing Done

- [x] successfully able to launch the driver in a ROS 2 environment and control the robot, call services, etc.