[SW-1121] Parameter cleanup in example yaml and spot_ros2 node (#407)
## Change Overview

There were a few parameters in `spot_ros_example.yaml` that are declared but completely unused. Given that we direct people to this as an example, I wanted to clean this up so that only relevant parameters remain. 

Additionally, this removes the image publishers from the main python `spot_ros2` node as they have been unused for a long time and were related to some of the parameters that were unused. 

The other config files also seem to be unmaintained but opted not to delete them as they might be used by some people? (planar.yaml, teleop.yaml, twist_mux.yaml) 

## Testing Done

- [x] Ran driver successfully
- [x] CI