[SW-736] Only sit on shutdown if driver holds the estop (#417)
## Change Overview

Currently, when you CTRL+C the driver, Spot will sit. If the estop is held by the tablet, this is annoying to deal with if you are starting and stopping the driver often, so this PR removes this feature.

If the estop is held by the driver, we still need to force sitting on shutdown, otherwise the robot will collapse to the ground if you CTRL+C.

EDIT: https://github.com/bdaiinstitute/spot_wrapper/pull/122 and https://github.com/bdaiinstitute/spot_ros2/pull/420 need to go in first so I can use the blocking version of the sit command. 

## Testing Done

- [x] started driver without an estop (i.e., estop is controlled by the tablet). Upon CTRL+C'ing the driver, the robot doesn't sit.
- [x] started driver with an estop (i.e., added `start_estop: True` to my config yaml). Robot cleanly sits down upon driver CTRL+C