Added dockerfile for running the driver (#415)
## Change Overview

I had spoken with @khughes-bdai on a previous Issue's discussion about creating a dockerfile that can be used to run the driver in production (whether that be on a computer or pushing it to Spot CORE I/O). This is separate from the Dockerfile found in the .devcontainer directory, as that is for creating a containerized development environment for something like GitHub Codespaces.

This PR adds a dockerfile to the root of the repository that can be used to run driver within a docker container. It uses the `osrf/ros-noetic-desktop-full` image as a base, and automatically downloads and installs the driver code from this repository.

I've pushed the built image to my docker hub account for convenience. It can be found at `ryroche/spot_ros2`

## Testing Done

This PR is a bit abnormal, as it does not modify any of the code. Thus I was unable to use traditional unit testing to ensure that the additions of this PR work.

- The docker image builds successfully
- Running the container, I was able to connect the driver to my robotics lab's spot robot and call the trigger services provided by the driver (`power_on`, `stand`, `sit`, etc.)
- Graphical tools such as `rviz2` work, you just need to enable your container to connect to your X11 server, whether that be by generating/exporting XAuth data to your container or by disabling authentication by running `xauth +x`