[SW-1322] Namespacing support for spot_ros2_control (#461)
This PR adds namespacing for the ros2 control nodes that get launched in order to mimic the setup we have in the spot_driver. There is a new launch argument `spot_name` in `spot_ros2_control.launch.py` that will apply a namespace to all of the nodes launched by this stack.

Discovered some fun issues with how ros2 control handles namespacing with respect to the config files for the controllers that get used. The names of the nodes in the config file must match the namespace applied. So a config file that looks like:
```
controller_manager:
  ros__parameters:
    update_rate: 333  # Hz

    joint_state_broadcaster:
      type: joint_state_broadcaster/JointStateBroadcaster
...
```
Does not actually work if you have launched the controller manager in a namespace, say `/spot/controller_manager`. Your config file needs to look like
```
spot/controller_manager:
  ros__parameters:
    update_rate: 333  # Hz

    joint_state_broadcaster:
      type: joint_state_broadcaster/JointStateBroadcaster
...
```
instead.

Putting all of these parameters in the global namespace is also not a good idea, as the config file is designed to hold many different types of controllers. (eg, you would also want to change `forward_position_controller` to `spot/forward_position_controller`, later on in this config, and this could cause issues if there were any parameters that had the same name between these two controllers). So the only solution is to generate a config file with the name appended to it at runtime, and use that. We already have done this in the spot_driver rviz launchfile [here](https://github.com/bdaiinstitute/spot_ros2/blob/main/spot_driver/launch/rviz.launch.py) so this follows that same convention about how to modify a template file and save it. Alongside this, we also have to do the same thing if we want rviz to show the robot in different namespaces too.

- [x] tested in mock mode: `ros2 launch spot_ros2_control spot_ros2_control.launch.py hardware_interface:=mock  mock_has_arm:=false spot_name:=test`
- [x] tested on robot: `ros2 launch spot_ros2_control spot_ros2_control.launch.py config_file:=spot_ros.yaml hardware_interface:=robot spot_name:=test`

Please create a checklist of tests you plan to do and check off the ones that have been completed successfully. Ensure that ROS 2 tests use `domain_coordinator` to prevent port conflicts. Further guidance for testing can be found on the [ros utilities wiki](https://github.com/bdaiinstitute/ros_utilities/wiki/Testing-guidelines).