[SW-1392] move all xacros to spot_description (#484)
## Change Overview

It was confusing to have two different xacros for spot (one in spot description and one in spot ros2 control). This PR moves everything into spot description. The default URDF that gets generated by the spot driver is still the same, but now if you add the `add_ros2_control_tags:=True` when generating `spot.xacro.urdf` we get the ROS 2 control tags added. 

Additionally fixes a one line issue with taking the lease in the hardware interface -- now even if the tablet has the lease, the hardware interface can take it. 

## Testing Done

- [x] successfully ran ros2 control stack on robot
- [x] successfully ran driver on robot