[SW-1394] add option for odom/fiducial/camera publishing from ros2 control launchfile (#485)
## Change Overview

Some nodes we have written in spot_driver are compatible with the ros2 control interface -- specifically, the state publisher node, the object sync node, and the camera publishers. This by default will launch these nodes when the ros2 control stack is run on robot. There is the option to turn off image publishing in case of bandwidth issues / if the user does not need it.

Since I was copying and pasting the launch arguments for the image publishers a lot, this PR also moves these into `spot_launch_helpers.py` so they can simply be imported into all the launchfiles that use it.

Finally, this changes the topic name that the low level joint states are published on from `/<robot>/joint_states` to `/<robot>/low_level/joint_states` and updates the nodes that rely on this accordingly. This is useful here because the state publisher node publishes to `/<robot>/joint_states` and we want to have a clear distinction between these two, as they are being published from two different sources via different BD API calls (unclear if the values the high / low level query produces are the same). 

## Testing Done

- [x] Tested wiggle arm example on robot to verify new joint states topic is working properly
- [x] Verified odom transforms and images are available on robot
- [x] TF tree is connected properly :

![frames_ros2control.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Xynj6CBpA3NqqBFfE8Q9/83fb0aef-6d03-4ab4-b156-e59cde5a5cff.png)