[SW-1483] Reorganize ROS 2 control packages (#511)
## Change Overview

This PR splits the hardware interface for Spot into its own package in an effort to clean up the ros2 control work and make things more modular. The ROS 2 control relevant packages in this repo now are:
* `spot_hardware_interface`: contains custom hardware interface plugin
* `spot_controllers`: contains custom controller plugin (and would contain any future controller plugins we may want to have in this repo)
* `spot_ros2_control`: contains the main launchfiles for bringing up the hardware interface + controllers via ros2 control, also contains the examples

As a part of this I also updated the READMEs in these packages and moved some utility functions out of a header and into a .cpp file. 


## Testing Done

- [x] tested examples on robot and in mock. The interface for bringing up the ros2 control stack + running the examples has stayed the same.