[N/A] better error message if robot is missing joint level api license (#548)
## Change Overview

This came up recently when running the ros2 control stack on a robot that had an expired joint level control license. The error message spammed in this state is just "Failed to get robot state" which is not very helpful. Now this tells the user to double check for the license. 

## Testing Done

N/A