[N/A] revert update_sigterm_sigkill_timeouts (#553)
## Change Overview

Reverts the calls to `update_sigterm_sigkill_timeout` introduced in https://github.com/bdaiinstitute/spot_ros2/pull/537. I have a very strong hunch that this is tanking CI execution time in our internal repo (after this change, we are only able to complete 17% of relevant tests before timing out, which is blocking my PR to bring in the latest changes from this repo).

## Testing Done

N/A