[SW-1727] improve spot ros2 readmes (#547)
## Change Overview

Readme's hadn't been updated in a while, so some info was old or not very helpful. This PR updates this info and splits out some info from the main readme (which was very long) into the relevant packages this repo contains. It also includes links to the wiki which has some more detailed information on what topics/actions/services are available. 

I also swapped out the image at the top of the main readme purely for ✨ aesthetics ✨ 

I welcome nitpicks as this is a very nitpicky PR :) 

## Testing Done

N/A