[SW-2049] ability to stream joint gains through hardware interface (#588)
## Change Overview

Adds k_q_p and k_qd_p to the command interface of Spot's hardware interface.These get set to an initial value via the `initial_value` field from the URDF, so existing logic here in the hardware interface to parse this input has been removed in favor of this new method. Existing controllers will not need to change anything, as they simply don't claim the interface. 

The following PR in the stack will set up a controller to allow users to easily change this at runtime.

This PR also depends on the changes from https://github.com/bdaiinstitute/spot_description/pull/3 to expose these new command interfaces in the URDF. 

## Testing Done
The default behavior from the user interface does not change with this PR.

- [x] Tested wiggle arm example successfully on robot with default gains
- [x] Tested wiggle arm example with different initial gains, verified that robot behavior adjusts accordingly
\\