[SW-2049] custom controller for streaming joint gains (#590)
## Change Overview

Adds a new controller, `SpotJointController` that lets you stream any combination of position, velocity, effort, k_q_p, and k_qd_p to Spot.
This is then integrated with `spot_ros2_control` as an optional controller that can be launched with the stack. This also includes documentation for this new controller and an example that can be run on robot.

## Testing Done

- [x] Tested new controller on robot, verified that differing gain values change robot motion.