[SW-2229] clean up autogeneration of ros2 control config files (#618)
## Change Overview

This PR cleans up the way our ros2 control configuration files are loaded for the spot ros2 control stack.

Configurations need to be namespaced appropriately to account for global namespaces and joint prefixes. The previous method involved reading from a template file, manually modifying elements to account for namespacing/joint prefixing, then writing to a new configuration file. This was difficult to read and annoying to expand to the template files (especially since there were two of them for arm + no arm robots). With this addition, we get rid of the steps that involved reading from a template file and just generate the whole thing in the launchfile based on input parameters (i.e. does the robot have an arm?). The template files are left as a reference but are no longer used.

## Testing Done

- [x] Ran stack, all controllers/broadcasters are still loaded in correctly for arm + no arm robots.

Signed-off-by: Mayank Mishra <mmayank74567@gmail.com>
Signed-off-by: kabirkedia <kabirkedia0111@gmail.com>