Example for mixed level API  (#711)
## Change Overview

I wanted to make an example for the mixed level API in this repo, as a more barebones example that doesn't use any of our internal tooling. 
As a part of this I updated the readme to point to this example and also added mypy checks for the `spot_ros2_control` package which adds a few minor changes to other files.
Arguably this example could also live in `spot_examples`, but I opted to put it in `spot_ros2_control/examples` to keep anything involving low level control isolated in this package. 

## Testing Done

- [x] Ran on robot